Implement SunnahTimes and fix tomorrow sunrise
This patch implements SunnahTimes. It also fixes a bug in how the
sunrise for the next day was calculated. Instead of computing tomorrow's
sunrise from the SolarTime for tomorrow, the code was just adding a day
to the previous day's SolarTime. This could cause a 1 minute offset in
certain cases. This patch fixes this problem.